+++ Source/SwiftLintCore/Extensions/Configuration+Parsing.swift
@@ -42,7 +42,7 @@ extension Configuration {
         dict: [String: Any],
         ruleList: RuleList = RuleRegistry.shared.list,
         enableAllRules: Bool = false,
-        onlyRule: String? = nil,
+        onlyRule: [String] = [],
         cachePath: String? = nil
     ) throws {
         func defaultStringArray(_ object: Any?) -> [String] { [String].array(of: object) ?? [] }
@@ -81,7 +81,7 @@ extension Configuration {
             analyzerRules: analyzerRules
         )
 
-        if onlyRule == nil {
+        if onlyRule.isEmpty {
             Self.validateConfiguredRulesAreEnabled(
                 parentConfiguration: parentConfiguration,
                 configurationDictionary: dict,
@@ -174,12 +174,12 @@ extension Configuration {
             }
 
             switch rulesMode {
-            case .allEnabled:
+            case .allCommandLine, .onlyCommandLine:
                 return
-            case .only(let onlyRules):
+            case .onlyConfiguration(let onlyRules):
                 let issue = validateConfiguredRuleIsEnabled(onlyRules: onlyRules, ruleType: ruleType)
                 issue?.print()
-            case let .default(disabled: disabledRules, optIn: optInRules):
+            case let .defaultConfiguration(disabled: disabledRules, optIn: optInRules):
                 let issue = validateConfiguredRuleIsEnabled(
                     parentConfiguration: parentConfiguration,
                     disabledRules: disabledRules,
@@ -201,9 +201,11 @@ extension Configuration {
         var disabledInParentRules: Set<String> = []
         var allEnabledRules: Set<String> = []
 
-        if case .only(let onlyRules) = parentConfiguration?.rulesMode {
+        if case .onlyConfiguration(let onlyRules) = parentConfiguration?.rulesMode {
             enabledInParentRules = onlyRules
-        } else if case .default(let parentDisabledRules, let parentOptInRules) = parentConfiguration?.rulesMode {
+        } else if case .defaultConfiguration(
+            let parentDisabledRules, let parentOptInRules
+        ) = parentConfiguration?.rulesMode {
             enabledInParentRules = parentOptInRules
             disabledInParentRules = parentDisabledRules
         }
@@ -243,7 +245,7 @@ extension Configuration {
         allEnabledRules: Set<String>,
         ruleType: any Rule.Type
     ) -> Issue? {
-        if case .allEnabled = parentConfiguration?.rulesMode {
+        if case .allCommandLine = parentConfiguration?.rulesMode {
             if disabledRules.contains(ruleType.identifier) {
                 return Issue.ruleDisabledInDisabledRules(ruleID: ruleType.identifier)
             }
@@ -264,7 +266,7 @@ extension Configuration {
                 if enabledInParentRules.union(optInRules).isDisjoint(with: allIdentifiers) {
                     return Issue.ruleNotEnabledInOptInRules(ruleID: ruleType.identifier)
                 }
-            } else if case .only(let enabledInParentRules) = parentConfiguration?.rulesMode,
+            } else if case .onlyConfiguration(let enabledInParentRules) = parentConfiguration?.rulesMode,
                       enabledInParentRules.isDisjoint(with: allIdentifiers) {
                 return Issue.ruleNotEnabledInParentOnlyRules(ruleID: ruleType.identifier)
             }